Fix unit tests race condition
There was a race condition between the commands executing the action and
out action spy assertions. We now make sure the execute function
completed before checking the spy.

This also has the nice consequence of forcing us to test the error
objects in the callbacks, as previously an error would have been
ignored.